move contentful client up to top level lib
